[Merge Gate Open Artifacts](2) Lift review artifact predicates into workflow-core
Moves isCurrentReviewGateArtifact, getCurrentReviewArtifacts,
getCurrentRequiredReviewArtifacts, and reviewGateIsApproved from
execution-engine into workflow-core, and re-exports them from
task-runner-review-gate so the public surface is unchanged.

Pure move. workflow-core cannot import execution-engine, so the orchestrator
needs these predicates on its own side of the dependency edge.
