+++ packages/workflow-core/src/index.ts
@@ -5,6 +5,7 @@ export * from './scheduler.js';
 export * from './orchestrator.js';
 export * from './task-id-scope.js';
 export * from './merge-conflict-error.js';
+export * from './review-gate-artifacts.js';
 export * from './plan-parser.js';
 export * from './fallback-policy.js';
 export * from './graph-mutation.js';
